server/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1

# LangChain 관련 패키지
langchain==0.3.19
langchain-core==0.3.40
langchain-anthropic==0.3.8
anthropic==0.48.0
jiter==0.8.2
langchain-community==0.3.18
langchain-text-splitters==0.3.6
pydantic-settings==2.8.1

# HuggingFace 관련 패키지 
torch-2.6.0
transformers-4.49.0